% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mult.R
\name{mult}
\alias{mult}
\title{Multiply together two numbers.}
\usage{
mult(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The product of \code{x} and \code{y}.
}
\description{
Multiply together two numbers.
}
\examples{
mult(1, 1)
mult(0.5, 0.26)
mult(exp(1), pi)
mult(-0.5, 1)
}

