% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_biden_word_proportion.R
\name{plot_biden_word_proportion}
\alias{plot_biden_word_proportion}
\title{An R function that produces a plot showing how the proportion of given words in Joe Biden's 2020 Speeches changed over time}
\usage{
plot_biden_word_proportion(words, datafile)
}
\arguments{
\item{words}{A character vector containing the word or words to be computed.}

\item{datafile}{The Dataset to be used in the function}
}
\value{
A plot/plots of the percentage change of words in Biden's speeches over time:
\describe{
\item{data_file}{The dataframe.}
\item{new}{A variable denoting the speech part number.}
\item{tokens}{Tokenizes speech and remove stop words.}
\item{tokens_group}{Groups tokenized words by speech number, computes proportion.}
\item{word_plot}{Plots word propoprtions over time with Linear Regression line.}
}
}
\description{
This function produces a plot showing the proportional change of words in Joe Biden's speeches over time.
First, the dialogue is tokenized and stop words are removed, then the proportion of the word or words is computed.
It also plots a scatter plot with a linear regression line using the \code{ggplot2} package.
}
\examples{
plot_biden_word_proportion(words= c("schools", "justice"), datafile=bidens_speeches_data)


}
\author{
10784424, 10782401, 10782689, 10757306.
}
