% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf-idf.R
\name{tf_idf}
\alias{tf_idf}
\title{An R function that computes the frequency of a term depending on how rarely it is used.
This function calculates the tf, idf and tf-idf indexes of words in the bidens_speeches_data dataset and extracts and plots the top 10 tf-idf words for each speech event.
First, the dataset is tokenized and stop words are removed, then the tf, idf and tf-idf indexes are computed.
It also produces a bar plot of the top ten  tf-idf words for each event in ascending order using the \code{ggplot2} package.}
\usage{
tf_idf(datafile)
}
\arguments{
\item{datafile}{The dataset containing words to be indexed.}
}
\value{
Plots of the tf-idf index of top 10 tf-idf words from each speech event:
\describe{
\item{data_file}{The dataframe.}
\item{df_tokenized}{A variable denoting the speech part number as well as tokenize speech.}
\item{df_token_cleaned}{Removes stop words.}
\item{df_final}{Groups tokenized words by location, computes tf-idf index.}
\item{top_words}{top 10 tf-idf words for each event.}
}
}
\description{
An R function that computes the frequency of a term depending on how rarely it is used.
This function calculates the tf, idf and tf-idf indexes of words in the bidens_speeches_data dataset and extracts and plots the top 10 tf-idf words for each speech event.
First, the dataset is tokenized and stop words are removed, then the tf, idf and tf-idf indexes are computed.
It also produces a bar plot of the top ten  tf-idf words for each event in ascending order using the \code{ggplot2} package.
}
\examples{
tf_idf(datafile = bidens_speeches_data)


}
\author{
10784424, 10782401, 10782689, 10757306.
}
