% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipf_law.R
\name{zipf}
\alias{zipf}
\title{An R function that produces regression statistics as well as a line plot illustration of zipf's law}
\usage{
zipf(datafile)
}
\arguments{
\item{datafile}{The dataset passed into the function containing text to be used.}
}
\value{
A plot/plots of the percentage change of words in Biden's speeches over time:
\describe{
\item{datafile}{The dataframe to be used.}
\item{df_tokenized}{A variable denoting the speech part number that also tokenizes speech.}
\item{word_group}{Groups tokenized speech by location and computes tf-idf index.}
\item{model}{The linear regression model of the independent and dependent variables.}
}
}
\description{
This function produces a plot showing the relationship between a word's rank in a document and its term frequency, in this case, the Joe Biden's dataset.
First, the dialogue is tokenized, then the term frequency (tf) of the words are computed alongside their ranking.
It produces a plot of the word rank and term frequency with a linear regression line using the \code{ggplot2} package.
}
\examples{
zipf(datafile=bidens_speeches_data)


}
\author{
10784424, 10782401, 10782689, 10757306.
}
